Upgrade near-sdk and borsh dependencies
Update near-sdk to 5.24 and borsh to 1.6 across main contract and v2
fixture. Migrate from deprecated collections module to store module
(IterableMap, LookupSet). Update type annotations for compatibility
with latest SDK (u32 for counts and pagination parameters).

All 51 integration tests pass.

+++ contracts/verified-accounts/src/interface.rs
@@ -248,7 +248,7 @@ pub trait VerifiedAccountsInterface {
     // ==================== Metadata ====================
 
     /// Get total number of verified accounts.
-    fn get_verified_count(&self) -> u64;
+    fn get_verified_count(&self) -> u32;
 
     /// Check if contract is paused.
     ///
